+++ R/runDebug.R
@@ -4,6 +4,7 @@
 #' @description runDebug is an internal function used by GRD and superb
 #'      to help in debugging the functions. It assigns in the global environment
 #'      the variables that are local to a function so that they become visible.
+#'      Use `options("superb.feedback" = "all")` to turn all debug on.
 #'
 #' @md
 #'