+++ tests/testthat/test_GRD.R
@@ -1,4 +1,4 @@
-context("Testing GRD")
+context("Testing GRD()")
 
 #################################################
 ##                                             ##