+++ vignettes/TheMakingOf.R
@@ -1,0 +1,9 @@
+## ----message=FALSE, warning=FALSE, echo=TRUE, eval=TRUE-----------------------
+## Load relevant packages
+library(superb)             # for superbPlot
+library(ggplot2)            # for all the graphic directives
+library(gridExtra)          # for grid.arrange
+
+## ----message=FALSE, echo=TRUE, eval=TRUE--------------------------------------
+head(dataFigure1)
+